WIP - Have icons in the outline + other fixes and adaptations

+++ plugins/com.aptana.editor.php/src/com/aptana/editor/php/internal/parser/nodes/PHPExtendsNode.java
@@ -72,6 +72,6 @@ public class PHPExtendsNode extends PHPBaseParseNode /* implements IResolvableIt
 			return children;
 		}
 
-		return new PHPBaseParseNode[] { new PHPBaseParseNode(0, 0, 0, 0, Messages.PHPExtendsNode_NonOnBuildPath0) };
+		return new PHPBaseParseNode[] { new PHPBaseParseNode((short)0, 0, 0, 0, Messages.PHPExtendsNode_NonOnBuildPath0) };
 	}
 }
